requirements.txt
Django==4.2.20
python-decouple==3.6
psycopg2-binary==2.8.6